Extra: now working with namespaces
Got the program to work with a testfile that used a namespace.

Im still working on how to best pass the data back down to the main program. Right now it's a terrible implementation, but it works!

Also, still working on finding out what's causing that double method printing bug